bpf: split off IPsec logic from encrypt.h
encrypt.h is starting to accumulate logic that is shared between IPsec and
Wireguard (fitting the file name). It's time to split off a dedicated
header file for IPsec, so that we can correctly manage CODEOWNERS
attribution for it.
